+++ lib/pages/intro/intro_page.dart
@@ -1,3 +1,11 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (c) 2019-2026 FluffyChat Contributors
+// Copyright (c) 2026 Simon
+//
+// MODIFICATIONS:
+// - 2026-02-07: Complete FluffyChat to FF Chat rebranding - Simon
+// - 2026-02-07: Adjust logo scaling - Simon
+
 import 'package:flutter/material.dart';
 
 import 'package:flutter_linkify/flutter_linkify.dart';
@@ -84,9 +92,15 @@ class IntroPage extends StatelessWidget {
                       padding: const EdgeInsets.symmetric(horizontal: 8.0),
                       child: Hero(
                         tag: 'info-logo',
-                        child: Image.asset(
-                          './assets/banner_transparent.png',
-                          fit: BoxFit.fitWidth,
+                        child: ConstrainedBox(
+                          constraints: const BoxConstraints(
+                            maxWidth: 200,
+                            maxHeight: 200,
+                          ),
+                          child: Image.asset(
+                            'assets/logo_transparent.png',
+                            filterQuality: FilterQuality.medium,
+                          ),
                         ),
                       ),
                     ),